project/requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
httplib2==0.19.1
idna==2.10
numpy==1.20.3
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.2.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
